requirements.txt
Django==1.8.3
django-colorful==1.1.0
psycopg2==2.6.1
djangorestframework==3.2.3
django-braces==1.8.1
Pillow==2.9.0